Break down spawn groups into smaller files, and export dino remaps
