+++ export/wiki/maps/gathering_complex.py
@@ -4,7 +4,7 @@ from typing import Any, Dict, Iterable, Optional, Set, Type, Union, cast
 from automate.hierarchy_exporter import ExportModel
 from export.wiki.consts import DAMAGE_TYPE_RADIATION_PKG
 from export.wiki.models import MinMaxRange
-from export.wiki.stage_spawn_groups import convert_single_class_swap
+from export.wiki.spawn_groups.structs import convert_single_class_swap
 from export.wiki.types import BiomeZoneVolume, DayCycleManager_Gen1, ExplorerNote, HexagonTradableOption, \
     MissionDispatcher_MultiUsePylon, NPCZoneManager, PrimalWorldSettings, SupplyCrateSpawningVolume, TogglePainVolume
 from ue.asset import ExportTableItem
