+++ export/wiki/stage_spawn_groups.py
@@ -1,44 +1,18 @@
-from collections import defaultdict
-from typing import Any, Dict, Iterable, List, Optional, cast
+from typing import Any, Dict, Optional, cast
 
-from automate.hierarchy_exporter import ExportFileModel, ExportModel, Field, JsonHierarchyExportStage
+from automate.hierarchy_exporter import ExportFileModel, ExportModel, JsonHierarchyExportStage
 from ue.asset import UAsset
 from ue.proxy import UEProxyStructure
-from ue.utils import sanitise_output
 
-from .maps.models import WeighedClassSwap
+from .spawn_groups.additions import RuntimeGroupAddition, segregate_container_additions
+from .spawn_groups.remaps import DinoRemap, convert_npc_remaps
+from .spawn_groups.structs import NpcGroup, NpcLimit, WeighedClassSwap, \
+    convert_class_swaps, convert_group_entry, convert_limit_entries
 from .types import NPCSpawnEntriesContainer
 
 __all__ = [
     'SpawnGroupStage',
 ]
-
-
-class Vector(ExportModel):
-    x: float
-    y: float
-    z: float
-
-
-class NpcEntry(ExportModel):
-    chance: float
-    bp: Optional[str]
-    offset: Vector = Vector(x=0, y=0, z=0)
-
-
-class NpcGroup(ExportModel):
-    name: str
-    weight: float
-    species: List[NpcEntry]
-    classSwaps: Optional[List[WeighedClassSwap]]
-
-
-class NpcLimit(ExportModel):
-    bp: str
-    mult: float = Field(..., title="Max desired NPC number multiplier")
-
-    class Config:
-        allow_population_by_field_name = True
 
 
 class NpcGroupsContainer(ExportModel):
@@ -48,16 +22,11 @@ class NpcGroupsContainer(ExportModel):
     limits: List[NpcLimit] = list()
 
 
-class RuntimeGroupAddition(ExportModel):
-    bp: Optional[str] = Field(...)
-    entries: List[NpcGroup] = list()
-    limits: List[NpcLimit] = list()
-
-
 class SpawnGroupsExportModel(ExportFileModel):
     spawngroups: List[NpcGroupsContainer]
     classSwaps: List[WeighedClassSwap]
     externalGroupChanges: List[RuntimeGroupAddition]
+    dinoRemaps: List[DinoRemap]
 
     class Config:
         title = "NPC spawning groups data for the Wiki"
@@ -86,29 +55,32 @@ class SpawnGroupStage(JsonHierarchyExportStage):
         if modid:
             mod_data = self.manager.arkman.getModData(modid)
             assert mod_data
+
             package = mod_data.get('package', None)
             if package:
                 pgd_asset = self.manager.loader[package]
-                result: Dict[str, Any] = dict()
-                # Export global NPC replacements and runtime group
-                # additions from the mod PGD.
-                class_swaps = list(convert_class_swaps(pgd_asset))
-                ext_group_changes = segregate_container_additions(pgd_asset)
-                if class_swaps:
-                    result['classSwaps'] = class_swaps
-                if ext_group_changes:
-                    result['externalGroupChanges'] = ext_group_changes
-                return result
+                return self._get_pgd_data(pgd_asset)
         else:
             pgd_asset = self.manager.loader['/Game/PrimalEarth/CoreBlueprints/BASE_PrimalGameData_BP']
-            result = dict()
-            # Export global NPC replacements from core PGD.
-            class_swaps = list(convert_class_swaps(pgd_asset))
-            if class_swaps:
-                result['classSwaps'] = class_swaps
-            return result
+            return self._get_pgd_data(pgd_asset)
 
         return None
+
+    def _get_pgd_data(self, pgd: UAsset) -> Dict[str, Any]:
+        v: Dict[str, Any] = dict()
+
+        class_swaps = list(convert_class_swaps(pgd))
+        ext_group_changes = segregate_container_additions(pgd)
+        npc_remaps = convert_npc_remaps(pgd)
+
+        if class_swaps:
+            v['classSwaps'] = class_swaps
+        if ext_group_changes:
+            v['externalGroupChanges'] = ext_group_changes
+        if npc_remaps:
+            v['dinoRemaps'] = npc_remaps
+
+        return v
 
     def extract(self, proxy: UEProxyStructure) -> Any:
         container: NPCSpawnEntriesContainer = cast(NPCSpawnEntriesContainer, proxy)
@@ -128,115 +100,3 @@ class SpawnGroupStage(JsonHierarchyExportStage):
             out.limits = list(convert_limit_entries(container.NPCSpawnLimits[0].values))
 
         return out
-
-
-def convert_group_entry(struct) -> NpcGroup:
-    d = struct.as_dict()
-    out = NpcGroup(
-        name=str(d['AnEntryName']),
-        weight=d['EntryWeight'],
-        species=list(),
-    )
-
-    # Export zipped NPC entries
-    chances = d['NPCsToSpawnPercentageChance'].values
-    offsets = d['NPCsSpawnOffsets'].values
-    for index, kls in enumerate(d['NPCsToSpawn'].values):
-        npc = NpcEntry(
-            chance=chances[index] if index < len(chances) else 1,
-            bp=sanitise_output(kls),
-            offset=sanitise_output(offsets[index] if index < len(offsets) else Vector(x=0, y=0, z=0)),
-        )
-        out.species.append(npc)
-
-    # Export local random class swaps if any exist
-    swaps = d['NPCRandomSpawnClassWeights'].values
-    if swaps:
-        out.classSwaps = [convert_single_class_swap(entry.as_dict()) for entry in swaps]
-
-    return out
-
-
-def convert_limit_entries(array) -> Iterable[NpcLimit]:
-    for entry in array:
-        d = entry.as_dict()
-        npc_class = sanitise_output(d['NPCClass'])
-
-        if npc_class:
-            yield NpcLimit(bp=npc_class, mult=d['MaxPercentageOfDesiredNumToAllow'])
-
-
-def convert_single_class_swap(d) -> WeighedClassSwap:
-    result = WeighedClassSwap(from_class=sanitise_output(d['FromClass']),
-                              exact=bool(d.get('bExactMatch', True)),
-                              to=sanitise_output(d['ToClasses']),
-                              weights=d['Weights'].values)
-
-    if d['ActiveEvent'] and d['ActiveEvent'].value and d['ActiveEvent'].value.value:
-        # Assigning "None" here is safe as it is the field default and therefore omitted
-        result.during = str(d['ActiveEvent'])
-
-    return result
-
-
-def convert_class_swaps(pgd: UAsset) -> Iterable[WeighedClassSwap]:
-    assert pgd.default_export
-    export_data = pgd.default_export.properties
-    d = export_data.get_property('GlobalNPCRandomSpawnClassWeights', fallback=None)
-    if not d:
-        return None
-
-    for entry in d.values:
-        yield convert_single_class_swap(entry.as_dict())
-
-
-def segregate_container_additions(pgd: UAsset) -> Optional[List[RuntimeGroupAddition]]:
-    if not pgd.default_export:
-        return None
-
-    export_data = pgd.default_export.properties
-    d = export_data.get_property('TheNPCSpawnEntriesContainerAdditions', fallback=None)
-    if not d:
-        return None
-
-    # Extract the addition entries
-    change_queues: Dict[str, List[RuntimeGroupAddition]] = defaultdict(list)
-    for add in d.values:
-        add = add.as_dict()
-        klass = add['SpawnEntriesContainerClass']
-        entries = add['AdditionalNPCSpawnEntries'].values
-        limits = add['AdditionalNPCSpawnLimits'].values
-        if not klass.value.value or (not entries and not limits):
-            continue
-
-        out = RuntimeGroupAddition(
-            bp=sanitise_output(klass),
-            entries=[convert_group_entry(entry) for entry in entries],
-            limits=convert_limit_entries(limits),
-        )
-
-        # Skip if no data
-        if not out.limits and not out.entries:
-            continue
-
-        # Append to the fragment list
-        klass_name = klass.format_for_json()
-        change_queues[klass_name].append(out)
-
-    # Merge
-    vs = []
-    for klass_name, changes in change_queues.items():
-        if len(changes) == 1:
-            vs.append(changes[0])
-            continue
-
-        out = changes.pop(0)
-
-        # Concat data arrays
-        for extra in changes:
-            out.entries += extra.entries
-            out.limits += extra.limits
-
-        vs.append(out)
-
-    return vs
